this is sorta working... but, when creating the initial hypercite (copy-hypercite) this isn't sent to server instantly, so, there is a strong chance it wont work quickly enough for most users... it's currently just there as an initial step.
> updateCitationForExistingHypercite begins its work. Crucially, where does it look for the original hypercite?
>
> - const nodeChunks = await getNodeChunksFromIndexedDB(booka);
> - const existingHypercite = await getHyperciteFromIndexedDB(booka, hyperciteIDa);

Now that we got the main thing syncing and what not... i want to return to this point, because, I realised something when I was reading this. At this point, where hyperciting logic looks to indexedDB for the hypercite record, this works great and quicly when user is in the same browser... but, there are use cases, for example, when user is on macbook, copies hypercite, and then pastes into their iphone... or, maybe they just use two different browsers for yolo reasons. so, it would be good to generate here a process whereby, if site detects a hypercite paste, but does not find the associated record within indexedDB, it will look for one in the backend server database, it will then proceed with the hyperciting, and what not, and if not, potentially we should alert the user that, we detected them pasting a hypercite, but we can't locate the record. potentially we would even offer them the option of writing, to explain, what kind of hyperciting operation they were attempting... or something... as this could be a critical information to get maxium data on. what do you think about this, and if we went forward with this, how much work would be involved, and so on, and what would the tradeoffs be?